requirements.txt
wget>=3.2
pyyaml>=5.4
pycryptodome>=3.9.7
pexpect>=4.7.0
